back-end/requirements.txt
fastapi
uvicorn
gunicorn  
sqlalchemy
psycopg2-binary 
python-dotenv 
python-jose[cryptography] 
passlib[bcrypt]
pwdlib
sendgrid  
line-bot-sdk  